Use exprType instead of SimpleNode for the node/expression in Use and Reference.
The one case where it wasn't an expression already was in visitImportFrom in DefinitionVisitor, for registering a NameUse for Class in "from module import Class" so that reference finders can find them (when renaming in IDE). But imports need to be treated specially for renaming anyway because there can't be a NameUse for module in "import pkg.module".